add files/directories_test.go; move tests
